refactor: BookQueryServiceの返却型をBookSummaryDtoに変更     - 著者IDに紐づく書籍一覧を取得するメソッドの返却型をBookDtoからBookSummaryDtoに変更し、冗長な著者情報を除外した
