+++ usecase/src/main/kotlin/com/bookmanagementsystem/usecase/author/book/read/ReadAuthorBookUsecase.kt
@@ -1,9 +1,9 @@
-package com.bookmanagementsystem.usecase.author.read
+package com.bookmanagementsystem.usecase.author.book.read
 
 import com.bookmanagementsystem.domain.author.Author
 import com.bookmanagementsystem.domain.author.AuthorRepository
 import com.bookmanagementsystem.domain.core.ID
-import com.bookmanagementsystem.usecase.book.BookDto
+import com.bookmanagementsystem.usecase.book.BookSummaryDto
 import com.bookmanagementsystem.usecase.book.read.BookQueryService
 import org.springframework.stereotype.Service
 
@@ -15,7 +15,7 @@ class ReadAuthorBookUsecase(
     /**
      * 指定された著者IDに紐づく書籍一覧を取得する
      */
-    fun execute(authorId: ID<Author>): List<BookDto> {
+    fun execute(authorId: ID<Author>): List<BookSummaryDto> {
         authorRepository.findById(authorId)
             ?: throw NoSuchElementException("著者が見つかりません。ID: ${authorId.value}")
