+++ usecase/src/main/kotlin/com/bookmanagementsystem/usecase/book/read/BookQueryService.kt
@@ -2,11 +2,11 @@ package com.bookmanagementsystem.usecase.book.read
 
 import com.bookmanagementsystem.domain.author.Author
 import com.bookmanagementsystem.domain.core.ID
-import com.bookmanagementsystem.usecase.book.BookDto
+import com.bookmanagementsystem.usecase.book.BookSummaryDto
 
 interface BookQueryService {
     /**
      * 著者IDに紐づく書籍一覧を取得する
      */
-    fun findByAuthorId(authorId: ID<Author>): List<BookDto>
+    fun findByAuthorId(authorId: ID<Author>): List<BookSummaryDto>
 }
